Address reconnect skeleton feedback
